test: move issue-534 profile preservation tests to new file for ratchet
